requirements.txt
bcrypt==4.1.2
fastapi==0.109.0
google-cloud-firestore==2.15.0
uvicorn==0.27.0
pydantic==2.6.0